chapter_003/src/test/java/ru/job4j/collections/sort/ListCompareTest.java
package ru.job4j.collections.sort;

import org.junit.Test;

import java.util.Arrays;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

/**
 * Test for ListCompare class
 */
public class ListCompareTest {
    /**
     * Test for ListCompare.compare method
     */
    @Test
    public void whenLeftAndRightEqualsThenZero() {
        ListCompare compare = new ListCompare();
        int rst = compare.compare(
                Arrays.asList(1, 2, 3),
                Arrays.asList(1, 2, 3)
        );
        assertThat(rst, is(0));
    }
    /**
     * Test for ListCompare.compare method
     */
    @Test
    public void whenLeftGreatRightThenMinus() {
        ListCompare compare = new ListCompare();
        int rst = compare.compare(
                Arrays.asList(1),
                Arrays.asList(1, 2, 3)
        );
        assertThat(rst, is(-1));
    }
    /**
     * Test for ListCompare.compare method
     */
    @Test
    public void whenLeftLessRightThenPlus() {
        ListCompare compare = new ListCompare();
        int rst = compare.compare(
                Arrays.asList(1, 2),
                Arrays.asList(1, 1)
        );
        assertThat(rst, is(1));
    }
}